add to skip existing computations

+++ data/create-static-page.py
@@ -11,12 +11,15 @@ import shutil
 
 
 def persist(output_dir, endpoint):
-    tira = Client()
     out = output_dir / Path(endpoint[1:])
     if endpoint.endswith('/'):
         out = out / 'index.json'
     out.parent.mkdir(exist_ok=True, parents=True)
 
+    if out.exists():
+        return json.load(open(out, 'r'))
+
+    tira = Client()
     response = tira.json_response(endpoint)
 
     with open(out, 'w') as f:
@@ -38,9 +41,13 @@ def main(output_dir: Path):
     extract_zip_from_url('https://github.com/tira-io/tira/releases/latest/download/frontend-build.zip', output_dir)
     shutil.copyfile(output_dir / "index.html", output_dir / "404.html")
 
+    persist(output_dir, "/api/role")
     tasks = persist(output_dir, "/api/task-list")['context']['task_list']
     for task in tqdm(tasks, 'Persist tasks'):
-        datasets = persist(output_dir, f'/api/datasets_by_task/{task["task_id"]}')
+        task_id = task["task_id"]
+        datasets = persist(output_dir, f'/api/datasets_by_task/{task_id}')
+        persist(output_dir, f'/api/task/{task_id}')
+        
 
 if __name__ == '__main__':
     main()
